change numerator. now it require less memory

+++ src/documents/Document.scala
@@ -8,8 +8,8 @@ import attribute.AttributeType
  * Date: 21.03.14
  * Time: 17:42
  */
-class Document(private val attributes: Map[AttributeType, Array[(Int, Int)]], val serialNumber: Int) {
-    def getAttributes(attributeType: AttributeType): Array[(Int, Int)] = attributes.getOrElse(attributeType, Array[(Int, Int)]())
+class Document(private val attributes: Map[AttributeType, Array[(Int, Short)]], val serialNumber: Int) {
+    def getAttributes(attributeType: AttributeType): Array[(Int, Short)] = attributes.getOrElse(attributeType, Array[(Int, Short)]())
 
     def contains(attribute: AttributeType) = attributes.contains(attribute)
